# Reinstate caching in schema generation