I spent a bit of time investigating the performance of `AIFunctionFactory.Create` where I discovered that schema generation is the most expensive factor by far is schema derivation. Even though the schema generation routines used to employ a caching mechanism in the past, this got removed during the recent refactorings since it no longer is necessary for leaf clients to resolve schemas on the fly.

This PR reinstates a different version of the caching mechanism for the purpose of speeding up `AIFunctionFactory.Create` when applied multiple times on the same delegate.

## Benchmarks

```C#
[MemoryDiagnoser]
public class Benchmark
{
    private static readonly Func<string, int, MyPoco?, MyPoco> _delegate = MyFunc;

    [Benchmark]
    public JsonElement CreateFunctionSchema() => AIJsonUtilities.CreateFunctionJsonSchema(_delegate.Method);

    [Benchmark]
    public JsonElement CreateFunctionSchema_CustomDescription() => AIJsonUtilities.CreateFunctionJsonSchema(_delegate.Method, description: "custom description");

    [Benchmark]
    public JsonElement CreateFunctionSchema_NewOptions() => AIJsonUtilities.CreateFunctionJsonSchema(_delegate.Method, serializerOptions: new(JsonSerializerOptions.Default));

    [Benchmark]
    public AIFunction CreateAIFunction() => AIFunctionFactory.Create(_delegate);

    [Benchmark]
    public AIFunction CreateAIFunction_CustomDescription() => AIFunctionFactory.Create(_delegate, description: "custom description");

    [Benchmark]
    public AIFunction CreateAIFunction_NewOptions() => AIFunctionFactory.Create(_delegate, serializerOptions: new(JsonSerializerOptions.Default));

    [Description("My awesome function")]
    public static MyPoco MyFunc([Description("Parameter 1")]string param1, int param2 = 42, MyPoco? other = null) => throw new NotImplementedException();

    public record MyPoco(int X, int Y);
}
```

### Main

| Method                                 | Mean     | Error     | StdDev    | Gen0   | Allocated |
|--------------------------------------- |---------:|----------:|----------:|-------:|----------:|
| CreateFunctionSchema                   | 6.345 us | 0.0848 us | 0.0708 us | 2.4109 |   9.95 KB |
| CreateFunctionSchema_CustomDescription | 6.041 us | 0.0357 us | 0.0298 us | 2.3499 |    9.7 KB |
| CreateFunctionSchema_NewOptions        | 6.438 us | 0.0260 us | 0.0230 us | 2.1973 |    9.1 KB |
| CreateAIFunction                       | 6.706 us | 0.0344 us | 0.0322 us | 2.5940 |  10.65 KB |
| CreateAIFunction_CustomDescription     | 6.298 us | 0.0393 us | 0.0368 us | 2.5330 |  10.45 KB |
| CreateAIFunction_CustomOptions         | 6.683 us | 0.0300 us | 0.0281 us | 2.3804 |   9.85 KB |


### PR

| Method                                 | Mean        | Error     | StdDev    | Gen0   | Allocated |
|--------------------------------------- |------------:|----------:|----------:|-------:|----------:|
| CreateFunctionSchema                   | 6,679.98 ns | 35.900 ns | 31.824 ns | 2.4109 |   10193 B |
| CreateFunctionSchema_CustomDescription | 6,126.56 ns | 43.308 ns | 38.391 ns | 2.3499 |    9929 B |
| CreateFunctionSchema_NewOptions        | 6,494.00 ns | 59.036 ns | 52.334 ns | 2.1973 |    9318 B |
| CreateAIFunction                       |    20.20 ns |  0.097 ns |  0.081 ns |      - |         - |
| CreateAIFunction_CustomDescription     |    28.19 ns |  0.182 ns |  0.171 ns | 0.0134 |      56 B |
| CreateAIFunction_NewOptions            | 7,309.19 ns | 72.739 ns | 64.481 ns | 2.5635 |   10854 B |



 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/extensions/pull/5908)