@eiriktsarpalis 2025-02-15
> We expect that to be something folks do, recreating an AIFunction for the same delegate over and over?

When working on the demo I felt enticed to incorporate the AIFunction declaration directly on the method like so:

```C#
chatOptions.Tools = [AIFunctionFactory.Create(GetNearestPorts)];
```

Which got me thinking that users would likely fall into the same performance trap.

> I'm wondering what the overhead of accessing the ConditionalWeakTable, accessing the ConcurrentDictionary, running the GetHashCode, etc. looks like when it doesn't pay dividends.

That scenario should be represented by the benchmarks creating a new JSO every time. As you can see, it does introduce performance regression compared to what we have in main today.