@eiriktsarpalis 2025-02-15
> > When working on the demo I felt enticed to incorporate the AIFunction declaration directly on the method like so:
> 
> Should we cache the AI Functions instead then?

I tried that and have a working prototype locally. What I did discover is that schema generation accounts for ~95% of the cost (as can be seen from the benchmark results) and caching on the `AIFunctionMetadata` level is trickier since there are more variables at play. It seemed like an obvious win to do it on the schema generation APIs instead.