@stephentoub 2025-02-15
Those numbers are very suspect to me, e.g.
```html
CreateAIFunction | 56.31 ns | 0.404 ns | 0.378 ns | 0.0172 | 72 B
```
That's suggesting with this change AIFunctionFactory.Create(delegate) only allocates 72 bytes and only took 56 nanoseconds? Without caching the AIFunction? It's doing a non-trivial amount of reflection. The AIFunction object alone must be close to 72 bytes, plus the delegates that get created for marshaling.

So I pulled down your change, and instead I see numbers like this:

| Method | Mean     | Error     | StdDev    | Allocated |
|------- |---------:|----------:|----------:|----------:|
| CreateAIFunction | 1.214 us | 0.0194 us | 0.0182 us |     816 B |