@eiriktsarpalis 2025-02-15
Here's updated numbers from a new run:

| Method                                 | Mean        | Error      | StdDev     | Gen0   | Gen1   | Allocated |
|--------------------------------------- |------------:|-----------:|-----------:|-------:|-------:|----------:|
| CreateFunctionSchema                   |    26.01 ns |   0.093 ns |   0.087 ns |      - |      - |         - |
| CreateFunctionSchema_CustomDescription |    38.41 ns |   0.172 ns |   0.161 ns |      - |      - |         - |
| CreateFunctionSchema_NewOptions        | 7,505.62 ns | 139.908 ns | 130.870 ns | 2.3804 |      - |   10167 B |
| CreateAIFunction                       |   612.79 ns |   5.975 ns |   5.589 ns | 0.1945 |      - |     816 B |
| CreateAIFunction_CustomDescription     |   224.36 ns |   1.081 ns |   0.903 ns | 0.1452 |      - |     608 B |
| CreateAIFunction_NewOptions            | 7,657.49 ns |  61.634 ns |  54.637 ns | 2.5635 | 0.0610 |   10779 B |

And here's the same benchmarks with additional caching implemented on the AIFunction layer:

| Method                                 | Mean        | Error      | StdDev     | Gen0   | Allocated |
|--------------------------------------- |------------:|-----------:|-----------:|-------:|----------:|
| CreateFunctionSchema                   |    27.15 ns |   0.085 ns |   0.079 ns |      - |         - |
| CreateFunctionSchema_CustomDescription |    32.35 ns |   0.215 ns |   0.201 ns |      - |         - |
| CreateFunctionSchema_NewOptions        | 7,648.95 ns | 140.465 ns | 196.912 ns | 2.4109 |   10167 B |
| CreateAIFunction                       |    60.11 ns |   0.307 ns |   0.287 ns | 0.0172 |      72 B |
| CreateAIFunction_CustomDescription     |    66.00 ns |   0.256 ns |   0.227 ns | 0.0305 |     128 B |
| CreateAIFunction_NewOptions            | 8,604.40 ns |  60.222 ns |  53.385 ns | 2.7466 |   11692 B |

So the numbers I quoted originally were most likely recorded with those changes accidentally unstashed. I've updated the OP to reflect the correct numbers.