@eiriktsarpalis 2025-02-18
@stephentoub as discussed, I pushed updates using caching on the AIFunctionFactory layer only. Here are the updated benchmark numbers:

| Method                                 | Mean        | Error     | StdDev    | Gen0   | Allocated |
|--------------------------------------- |------------:|----------:|----------:|-------:|----------:|
| CreateFunctionSchema                   | 6,679.98 ns | 35.900 ns | 31.824 ns | 2.4109 |   10193 B |
| CreateFunctionSchema_CustomDescription | 6,126.56 ns | 43.308 ns | 38.391 ns | 2.3499 |    9929 B |
| CreateFunctionSchema_NewOptions        | 6,494.00 ns | 59.036 ns | 52.334 ns | 2.1973 |    9318 B |
| CreateAIFunction                       |    20.20 ns |  0.097 ns |  0.081 ns |      - |         - |
| CreateAIFunction_CustomDescription     |    28.19 ns |  0.182 ns |  0.171 ns | 0.0134 |      56 B |
| CreateAIFunction_NewOptions            | 7,309.19 ns | 72.739 ns | 64.481 ns | 2.5635 |   10854 B |